Public YamlMappingFix
